Payment and results implementation

+++ Domain/DTOs/Requests/CreateRequestDtos/CreateExaminationRequestDto.cs
@@ -7,7 +7,7 @@ namespace App.Core.DTOs.Requests.CreateRequestDtos
         public required string ExamTitle { get; set; }
         public DateTime ExamDateAndTime { get; set; }
         public DateOnly ExamYear { get; set; }
-        public required decimal Fee { get; set; }
+        public required int Fee { get; set; }
         public required Guid CourseId { get; set; }
     }
 }
