+++ Domain/Entities/Examination.cs
@@ -7,7 +7,7 @@
         public DateOnly ExamYear { get; set; }
         public required Guid CourseId { get; set; }
         public required Course Course { get; set; }
-        public required decimal Fee { get; set; }
+        public required int Fee { get; set; }
         public ICollection<Result> Results { get; set; } = new List<Result>();
         public ICollection<UserExaminations> Examinations { get; set; } = new List<UserExaminations>();
     }
